+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Variable `ship-mate-environment` that can be set to prefill
+  `compilation-environment`.
+
 ## [v0.1.0]
 
 Initial release of the package that was spliced out of my own